Move TestSetupManager into a separate package
